+++ src/types/Settings.ts
@@ -43,6 +43,14 @@ type Settings = {
   };
   version: string;
   environment?: string;
+  influxDB: {
+    url: string;
+    org: string;
+    username: string;
+    password: string;
+    bucket: string;
+    token: string;
+  };
 };
 
 export default Settings;