+++ src/types/analytics/UsageData.ts
@@ -1,5 +1,0 @@
-export default interface UsageData {
-  apiKey: string;
-  route: string;
-  time?: string;
-}